Move service scenario to service data class.
